+++ src/main/java/nextstep/auth/AuthConfig.java
@@ -2,11 +2,11 @@ package nextstep.auth;
 
 import nextstep.auth.authentication.BasicAuthenticationFilter;
 import nextstep.auth.authentication.BearerTokenAuthenticationFilter;
-import nextstep.auth.authentication.filter.processing.UsernamePasswordAuthenticationFilter;
+import nextstep.auth.authentication.filter.processing.UsernamePasswordAuthenticationProcessingFilter;
 import nextstep.auth.authorization.AuthenticationPrincipalArgumentResolver;
 import nextstep.auth.context.SecurityContextPersistenceFilter;
 import nextstep.auth.token.JwtTokenProvider;
-import nextstep.auth.authentication.filter.processing.TokenAuthenticationInterceptor;
+import nextstep.auth.authentication.filter.processing.TokenAuthenticationProcessingFilter;
 import nextstep.member.application.LoginMemberService;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
@@ -27,8 +27,8 @@ public class AuthConfig implements WebMvcConfigurer {
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
         registry.addInterceptor(new SecurityContextPersistenceFilter());
-        registry.addInterceptor(new UsernamePasswordAuthenticationFilter(loginMemberService)).addPathPatterns("/login/form");
-        registry.addInterceptor(new TokenAuthenticationInterceptor(loginMemberService, jwtTokenProvider)).addPathPatterns("/login/token");
+        registry.addInterceptor(new UsernamePasswordAuthenticationProcessingFilter(loginMemberService)).addPathPatterns("/login/form");
+        registry.addInterceptor(new TokenAuthenticationProcessingFilter(loginMemberService, jwtTokenProvider)).addPathPatterns("/login/token");
         registry.addInterceptor(new BasicAuthenticationFilter(loginMemberService));
         registry.addInterceptor(new BearerTokenAuthenticationFilter(jwtTokenProvider));
     }
