+++ src/test/java/nextstep/auth/authentication/filter/processing/UsernamePasswordAuthenticationProcessingFilterTest.java
@@ -1,12 +1,8 @@
 package nextstep.auth.authentication.filter.processing;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import nextstep.auth.authentication.AuthenticationToken;
 import nextstep.auth.context.Authentication;
 import nextstep.auth.context.SecurityContextHolder;
-import nextstep.auth.token.JwtTokenProvider;
-import nextstep.auth.token.TokenRequest;
-import nextstep.auth.token.TokenResponse;
 import nextstep.member.application.LoginMemberService;
 import nextstep.member.domain.LoginMember;
 import org.junit.jupiter.api.BeforeEach;
@@ -22,26 +18,24 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-class TokenAuthenticationInterceptorTest {
+class UsernamePasswordAuthenticationProcessingFilterTest {
 
-    private static final String EMAIL = "email@email.com";
+    private static final String EMAIL = "admin@email.com";
     private static final String PASSWORD = "password";
-    public static final String JWT_TOKEN = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIiLCJuYW1lIjoiSm9obiBEb2UiLCJpYXQiOjE1MTYyMzkwMjJ9.ih1aovtQShabQ7l0cINw4k1fagApg3qLWiB8Kt59Lno";
 
-    private TokenAuthenticationInterceptor interceptor;
+    private UsernamePasswordAuthenticationProcessingFilter filter;
     private LoginMemberService loginMemberService;
-    private JwtTokenProvider jwtTokenProvider;
 
     @BeforeEach
     void setUp() {
         loginMemberService = mock(LoginMemberService.class);
-        jwtTokenProvider = mock(JwtTokenProvider.class);
-        interceptor = new TokenAuthenticationInterceptor(loginMemberService, jwtTokenProvider);
+        filter = new UsernamePasswordAuthenticationProcessingFilter(loginMemberService);
     }
+
 
     @Test
     void convert() throws IOException {
-        var authenticationToken = interceptor.convert(createMockRequest());
+        var authenticationToken = filter.convert(createMockRequest());
 
         assertAll(
                 () -> assertThat(authenticationToken.getPrincipal()).isEqualTo(EMAIL),
@@ -54,32 +48,29 @@ class TokenAuthenticationInterceptorTest {
         when(loginMemberService.loadUserByUsername(EMAIL))
                 .thenReturn(new LoginMember(EMAIL, PASSWORD, List.of("ROLE_ADMIN")));
 
-        var authentication = interceptor.authenticate(new AuthenticationToken(EMAIL, PASSWORD));
+        var authentication = filter.authenticate(new AuthenticationToken(EMAIL, PASSWORD));
+
 
         assertAll(
                 () -> assertThat(authentication.getPrincipal()).isEqualTo(EMAIL),
                 () -> assertThat(authentication.getAuthorities()).containsExactly("ROLE_ADMIN")
         );
-
     }
 
     @Test
-    void processing() throws Exception {
-        var roles = List.of("ROLE_ADMIN");
-        var response = new MockHttpServletResponse();
-        when(jwtTokenProvider.createToken(EMAIL, roles)).thenReturn(JWT_TOKEN);
+    void processing() {
+        filter.processing(new Authentication(EMAIL, List.of("ROLE_ADMIN")), new MockHttpServletResponse());
 
-        interceptor.processing(new Authentication(EMAIL, List.of("ROLE_ADMIN")), response);
-
-        var expectedResponse = new ObjectMapper().writeValueAsString(new TokenResponse(JWT_TOKEN));
-        System.out.println(expectedResponse);
-        assertThat(response.getContentAsString()).isEqualTo(expectedResponse);
+        var authentication = SecurityContextHolder.getContext().getAuthentication();
+        assertThat(authentication.getPrincipal()).isEqualTo(EMAIL);
+        assertThat(authentication.getAuthorities()).containsExactly("ROLE_ADMIN");
     }
 
     private MockHttpServletRequest createMockRequest() throws IOException {
         MockHttpServletRequest request = new MockHttpServletRequest();
-        TokenRequest tokenRequest = new TokenRequest(EMAIL, PASSWORD);
-        request.setContent(new ObjectMapper().writeValueAsString(tokenRequest).getBytes());
+        request.setParameter("username", EMAIL);
+        request.setParameter("password", PASSWORD);
         return request;
     }
+
 }
